Add `error_nodes` dictionary
This will helps us to keep track of the error nodes in the diagram. This
way, we will not find ourselves in a situation where we have to search
for the error nodes in the diagram or creating a new instance of the
error node every time we need to add a new error node.